Make RA acceleration values robot specific
Move setting of the parameters to SDK so we can have
different acceleration and deceleration values for
RA motion generation on different robot platforms.
This is required after tests on Kuka platform which
needs slower acceleration due to joint speed limits
on the kuka controller.

Default values are the ones tested before on URs and
Fanucs.